Edit expired_posts task to not delete posts with recent matches.

+++ lib/tasks/expired_posts.rake
@@ -3,6 +3,6 @@ task :expired_posts => :environment do
 
   expired_posts = Post.where('expiration < ?', Date.today)
   expired_posts.each do |expired_post|
-    expired_post.destroy
+    expired_post.destroy unless expired_post.has_recent_matches?
   end
 end
